backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
httpx
requests
openai

# Audio processing libraries
librosa
soundfile
pydub
numpy
scipy

# For visualization
matplotlib# Auth
entication
PyJWT# Adv
anced audio processing
pedalboard
pydub
ffmpeg-python